requirements.txt
fastapi[standard]
langchain
openai
celery
langchain-openai
python-dotenv